Moved classes related to Interpro management into another package
