+++ src/main/java/sttDB/service/interproServices/LineItems.java
@@ -1,4 +1,4 @@
-package sttDB.service;
+package sttDB.service.interproServices;
 
 public class LineItems {
